requirements.txt
# Django và DRF
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Swagger UI
drf-yasg==1.21.7

# Database
mysqlclient==2.2.0

# Celery và Redis
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
redis==5.0.1

# Cron jobs
django-crontab==0.7.1

# CSV Processing
pandas==2.1.3
numpy==1.25.2

# File handling
Pillow==10.1.0

# Environment variables
python-decouple==3.8

# Production server
gunicorn==21.2.0

# Development tools
django-extensions==3.2.3
